[SCRUM-137]:모바일 반응형 디자인 추가
