+++ package.json
@@ -11,10 +11,12 @@
   },
   "dependencies": {
     "@tailwindcss/vite": "^4.1.11",
+    "@types/react-responsive": "^8.0.8",
     "axios": "^1.10.0",
     "jwt-decode": "^4.0.0",
     "react": "^19.1.0",
     "react-dom": "^19.1.0",
+    "react-responsive": "^10.0.1",
     "react-router-dom": "^7.6.3",
     "react-toastify": "^11.0.5",
     "socket.io-client": "^4.8.1",
