+++ src/components/chat/ChatAPI.tsx
@@ -12,6 +12,7 @@ export const chatService = {
       const response = await apiClient.get('/group/init/chat', {
         params: { canvas_id: canvasId },
       });
+      console.log(response);
       const { defaultGroupId, groups, messages } = response.data.data;
       // 메시지를 시간순으로 정렬 (오래된 것부터)
       const sortedMessages = messages.sort(
